Refactor favourite button folder
Rename folder to be more consistent with rest of app.
